+++ tests/unit/helpers_test.sh
@@ -227,3 +227,14 @@ EOF
   assert_equals "0.10.1" "$(helpers::get_latest_tag)"
   unset -f git # remove the mock
 }
+
+function test_to_run_with_filter_matching_string_in_function_name() {
+  local functions=("test_my_awesome_function" "test_your_awesome_function" "test_so_lala_function")
+
+  assert_equals\
+    "test_your_awesome_function" "$(helper::get_functions_to_run "test" "test_your_awesome_function" "${functions[*]}")"
+
+  assert_equals\
+    "test_my_awesome_function test_your_awesome_function"\
+    "$(helper::get_functions_to_run "test" "awesome" "${functions[*]}")"
+}